requirements.txt
pbr>=1.1.0
virtualenv

github3.py>=1.1.0
PyYAML>=3.1.0
paramiko>=2.0.1
GitPython>=2.1.8
python-daemon>=2.0.4,<2.1.0
extras
statsd>=3.0
voluptuous>=0.10.2
gear>=0.13.0,<1.0.0
apscheduler>=3.0
PrettyTable>=0.6,<0.8
babel>=1.0
netaddr
kazoo
sqlalchemy
alembic
cryptography>=1.6
cachecontrol
cachetools
pyjwt
iso8601
psutil
fb-re2>=1.0.6
paho-mqtt
cherrypy
ws4py
routes
slackclient==1.3.1
